rls is now enforced on products table
