+++ prisma/schema.prisma
@@ -48,8 +48,8 @@ model Products {
     stockAvailable  Int
     description     String?
     sales           Sales[]
-    Organizations   Organizations @relation(fields: [organizationsId], references: [id])
-    organizationsId String
+    Organizations   Organizations @relation(fields: [organization_id], references: [id])
+    organization_id String
 
     // @@schema("auth")
 }
