+++ src/server/api/root.ts
@@ -14,7 +14,7 @@ import { accounting } from "./routers/expenses";
  * All routers added in /api/routers should be manually added here.
  */
 export const appRouter = createTRPCRouter({
-  // products,
+  products,
   accounting,
   // supplier,
   // invoices,
