+++ src/server/api/routers/products.ts
@@ -1,19 +1,13 @@
 import { createTRPCRouter, protectedProcedure } from "~/server/api/trpc";
 import z from "zod";
-import { productSchema } from "~/app/dashboard/products/new/page";
-import {
-  FAILED_TO_CREATE,
-  NOT_FOUND,
-  organizationEmailSchema,
-} from "~/utils/constants";
-import useOrganizationId from "~/utils/hooks/useOrganizationId";
+import { productSchema } from "~/app/dashboard/products/_components/schema";
+import { FAILED_TO_CREATE, NOT_FOUND } from "~/utils/constants";
 
 export const products = createTRPCRouter({
   add: protectedProcedure
     .input(productSchema)
     .mutation(async ({ input, ctx }) => {
       try {
-        const organizationId = "";
         const newProduct = await ctx.db.products.create({
           data: {
             name: input.name,
@@ -21,7 +15,8 @@ export const products = createTRPCRouter({
             sellingPrice: input.sellingPrice,
             stockAvailable: input.stockAvailable,
             description: input.description,
-            organizationsId: organizationId,
+            // @ts-ignore
+            organization_id: ctx.user?.id,
           },
         });
         return newProduct;
@@ -32,12 +27,15 @@ export const products = createTRPCRouter({
     }),
 
   all: protectedProcedure
-    .input(organizationEmailSchema)
-    .query(async ({ ctx, input }) => {
+    .query(async ({ ctx }) => {
       try {
         return await ctx.db.products.findMany({
           orderBy: {
             createdAt: "desc",
+          },
+          where: {
+            // @ts-ignore
+            organization_id: ctx.user?.id,
           },
         });
       } catch (cause) {
